A paper by Biogen
The clinical data research network (CDRN) is designed to gather real-world evidence that, in combination with randomized clinical trials, will help Biogen better characterize the safety and effectiveness of medicines for conditions like Alzheimer’s disease.

While randomized controlled trials are the "gold standard" when trying to determine a drug's efficacy, they often seek homogeneity of the patient population to avoid variance introduced by patient and disease characteristics. This eliminates bias and allows researchers to better examine cause-effect relationships between a treatment and health outcomes. 

But that means the results may not be generalizable to a target patient population, Parks said. That's why Biogen is turning to real-world evidence for better insights and knowledge on how medicines perform, especially in the case of a novel medicine or new class of therapies.